snapshotting now working for imperative searn

+++ vowpalwabbit/searn_sequencetask.h
@@ -29,7 +29,7 @@ namespace SequenceTask {
 }
 
 namespace SequenceTask_Easy {
-  void initialize(vw&, uint32_t&);
+  void initialize(vw&, size_t&);
   void finish(vw&);
   void structured_predict_v1(vw&, ImperativeSearn::searn&, example**,size_t,stringstream*,stringstream*);
 }
